+++ functions.h
@@ -2,4 +2,4 @@
  * Define the prototypes of functions defined in functions.c
  */
 
-long double D(double k_perp, double omega);
+double D(double k_perp, double omega);
